requirement.txt
streamlit
tensorflow
torch
Pillow
numpy
matpotlib
seaborn
scikit-learn